refactor: file structure for combat folder
