+++ src/components/charsheet/combat/defense/ArmorHelpContent.tsx
