+++ src/components/charsheet/combat/wounds/StabilizedControl.tsx
