+++ src/components/charsheet/combat/wounds/WoundBox.tsx
