+++ src/components/charsheet/combat/wounds/WoundHelpContent.tsx
