+++ src/components/charsheet/common/topbar/StatsStrip.tsx
