+++ src/components/charsheet/common/topbar/WoundIndicator.tsx
