+++ src/components/charsheet/dossier/StatColumn.tsx
