+++ src/components/charsheet/dossier/StatsPanel.tsx
